backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.1
pandas==2.1.3
PyGithub==2.1.1
python-multipart==0.0.6
httpx==0.25.2
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0